Terminatorで実行可能なように拡張

+++ RUN-DOCKER-CONTAINER.sh
@@ -5,63 +5,51 @@
 #   Setup Variable
 #
 # ==================================================================================================================
-export HOST_UID=$(id -u)
-export HOST_GID=$(id -g)
-export APP_USER_NAME=app
-export SIMULATOR_USER_NAME=simulator
+export HOST_UID="$(id -u)"
+export HOST_GID="$(id -g)"
+export APP_USER_NAME="app"
+export SIMULATOR_USER_NAME="simulator"
 #export APP_USER_NAME=${USER}
 
-app_container=emlab-tutorial-client
-simulation_container=emlab-tutorial-simulator
-N_SERVICE=2
-
-if [ -e /proc/driver/nvidia/version ]; then
-  export DOCKER_RUNTIME=nvidia
-else
-  export DOCKER_RUNTIME=runc
-fi
+readonly APP_CONTAINER="emlab-tutorial-client"
+readonly SIMULATION_CONTAINER="emlab-tutorial-simulator"
+readonly N_SERVICE=2
+readonly LOCK_FILE="/tmp/catkin.lock"
+readonly COMPLETE_WORD="Complete initialization of entrypoint"
+readonly COMPOSE_FILE="./docker/docker-compose.yml"
+readonly PIPE=/tmp/entrypoint_monitoring.pipe
 
 # ==================================================================================================================
 #
-#   Run container
+#   [Function] Run container
 #
 # ==================================================================================================================
-LOCK_FILE=/tmp/catkin.lock
-if [ "$(docker inspect --format="{{.State.Status}}" ${app_container})" != "running" ]; then
-  # Generate lock file
-  if [ ! -e ${LOCK_FILE} ]; then
-    touch ${LOCK_FILE}
-  fi
-
+function compose_up() {
   # docker-compose up
-  cd $(dirname "$0")
-  docker-compose -f ./docker/docker-compose.yml up -d
-  compose_status=$?
-  if [ ${compose_status} != 0 ]; then
-    rm ${LOCK_FILE}
-    exit 1
-  fi
+  docker-compose -f "${COMPOSE_FILE}" up -d
+  return $?
+}
 
-  # ==================================================================================================================
-  #
-  #   Wait for entrypoint
-  #
-  # ==================================================================================================================
-  CompleteWord="Complete initialization of entrypoint"
-
-  PIPE=/tmp/entrypoint_monitoring.pipe
+# ==================================================================================================================
+#
+#   [Function] Wait for entrypoint
+#
+# ==================================================================================================================
+function wait_for_entrypoint() {
   if [ -e ${PIPE} ]; then
     rm ${PIPE}
   fi
   mkfifo ${PIPE}
-  docker-compose -f ./docker/docker-compose.yml logs -f --since 0s | tee ${PIPE} >/dev/null &
-  PID=$!
-  count=0
 
+  docker-compose -f "${COMPOSE_FILE}" logs -f --since 0s | tee ${PIPE} >/dev/null &
+  PID=$!
+
+  local count=0
+  local line
   cat ${PIPE} | while read line; do
     echo ${line}
-    if [[ ${line} =~ ${CompleteWord} ]]; then
-      count=$((count+1))
+    if [[ ${line} =~ ${COMPLETE_WORD} ]]; then
+      count=$((count + 1))
     fi
     if [[ ${count} == ${N_SERVICE} ]]; then
       kill $PID
@@ -70,30 +58,50 @@ if [ "$(docker inspect --format="{{.State.Status}}" ${app_container})" != "runni
   done
 
   rm ${PIPE}
+}
 
-  docker exec -it -u ${APP_USER_NAME} ${app_container} /bin/bash -i -c "source /opt/ros/noetic/setup.bash; cd ~/catkin_ws; catkin build"
-  docker exec -it -u ${SIMULATOR_USER_NAME} ${simulation_container} /bin/bash -i -c "source /opt/ros/noetic/setup.bash; cd ~/catkin_ws; catkin build"
-  rm ${LOCK_FILE}
+# ==================================================================================================================
+#
+#   [Function] catkin build
+#
+# ==================================================================================================================
+function catkin_build() {
+  docker exec -it -u ${APP_USER_NAME} ${APP_CONTAINER} /bin/bash -i -c "source /opt/ros/noetic/setup.bash; cd ~/catkin_ws; catkin build"
+  docker exec -it -u ${SIMULATOR_USER_NAME} ${SIMULATION_CONTAINER} /bin/bash -i -c "source /opt/ros/noetic/setup.bash; cd ~/catkin_ws; catkin build"
+}
 
-else
-  # ==================================================================================================================
-  #
-  #   Wait docker-compose up
-  #
-  # ==================================================================================================================
-  if [ -e ${LOCK_FILE} ]; then
-    echo "Wait for docker-compose up ..."
+# ==================================================================================================================
+#
+#   main
+#
+# ==================================================================================================================
+cd $(dirname "$0")
+if [ "$(docker inspect --format="{{.State.Status}}" ${APP_CONTAINER})" != "running" ]; then
+    # Generate lock file
+  if [ ! -e ${LOCK_FILE} ]; then
+    touch ${LOCK_FILE}
   fi
 
-  while [ -e ${LOCK_FILE} ]; do
-    sleep 0.1
-  done
+  compose_up
+  if [ $? != 0 ]; then
+    rm ${LOCK_FILE}
+    exit 1
+  fi
 
+  wait_for_entrypoint
+  catkin_build
+
+  rm ${LOCK_FILE}
+
+  exit 0
 fi
 
-# ==================================================================================================================
-#
-#   Enter the docker container
-#
-# ==================================================================================================================
-#docker exec -it -u ${APP_USER_NAME} ${app_container} /bin/bash
+if [ ! -e ${LOCK_FILE} ]; then
+  echo "Containers are already activated."
+  exit 0
+fi
+
+echo "Wait for docker-compose up ..."
+while [ -e ${LOCK_FILE} ]; do
+  sleep 0.1
+done
